Update solution - CodeHub

+++ 8-kyu/find-the-smallest-integer-in-the-array/find-the-smallest-integer-in-the-array.java
@@ -1,5 +1,5 @@
 public class SmallestIntegerFinder {
     public static int findSmallestInt(int[] args) {
-        return (int) Math.min(args);
+        return Math.min(args);
     }
 }
